+++ src/main/java/lotto/common/LottoResult.java
