+++ src/main/java/lotto/common/exception/InvalidInputException.java
@@ -1,0 +1,2 @@
+package lotto.common.exception;public class InvalidInputException {
+}
