+++ src/main/java/lotto/model/generator/LottoGenerator.java
@@ -1,0 +1,2 @@
+package lotto.model.generator;public interface LottoGenerator {
+}
