+++ src/main/java/lotto/model/statistic/LottoStatistic.java
@@ -1,0 +1,2 @@
+package lotto.model;public class LottoStatistic {
+}
